+++ app/src/main/java/dev/leonlatsch/photok/other/Utils.kt
@@ -1,5 +1,5 @@
 /*
- *   Copyright 2020 Leon Latsch
+ *   Copyright 2020-2021 Leon Latsch
  *
  *   Licensed under the Apache License, Version 2.0 (the "License");
  *   you may not use this file except in compliance with the License.
@@ -19,16 +19,16 @@ package dev.leonlatsch.photok.other
 import android.content.ContentResolver
 import android.content.Context
 import android.content.Intent
+import android.graphics.Bitmap
+import android.graphics.Matrix
 import android.net.Uri
 import android.os.Handler
 import android.os.Looper
 import android.provider.MediaStore
-import android.view.View
-import android.view.Window
 import androidx.appcompat.app.AppCompatDelegate
-import androidx.fragment.app.Fragment
-import dev.leonlatsch.photok.BaseApplication
-import dev.leonlatsch.photok.ui.StartActivity
+import androidx.exifinterface.media.ExifInterface
+import timber.log.Timber
+import java.io.ByteArrayInputStream
 
 /**
  * Get a file's name.
@@ -51,42 +51,9 @@ fun getFileName(contentResolver: ContentResolver, uri: Uri): String? {
 }
 
 /**
- * Toggle the system ui.
- * Used for fullscreen.
- */
-fun toggleSystemUI(window: Window?) {
-    window ?: return
-    val uiOptions: Int = window.decorView.systemUiVisibility
-    var newUiOptions = uiOptions
-
-    newUiOptions = newUiOptions xor View.SYSTEM_UI_FLAG_HIDE_NAVIGATION
-    newUiOptions = newUiOptions xor View.SYSTEM_UI_FLAG_FULLSCREEN
-    newUiOptions = newUiOptions xor View.SYSTEM_UI_FLAG_IMMERSIVE
-
-    window.decorView.systemUiVisibility = newUiOptions
-}
-
-/**
- * Restart the app. Clear back stack and start [StartActivity]
- */
-fun restartAppLifecycle(context: Context) {
-    val intent = Intent(context, StartActivity::class.java)
-    intent.flags = Intent.FLAG_ACTIVITY_NEW_TASK or Intent.FLAG_ACTIVITY_CLEAR_TASK
-    context.startActivity(intent)
-}
-
-/**
  * Post a [operation] to the main looper.
  */
 fun runOnMain(operation: () -> Unit) = Handler(Looper.getMainLooper()).post(operation)
-
-/**
- * Extension for starting an activity for result and disable lock timer in [BaseApplication].
- */
-fun Fragment.startActivityForResultAndIgnoreTimer(intent: Intent, reqCode: Int) {
-    startActivityForResult(intent, reqCode)
-    BaseApplication.ignoreNextTimeout()
-}
 
 /**
  * Update the app design.
@@ -103,4 +70,62 @@ fun setAppDesign(design: String?) {
 
     nightMode ?: return
     AppCompatDelegate.setDefaultNightMode(nightMode)
+}
+
+fun openUrl(context: Context, url: String?) {
+    url ?: return
+    val intent = Intent(Intent.ACTION_VIEW)
+    intent.data = Uri.parse(url)
+    context.startActivity(intent)
+}
+
+/**
+ * Reset all orientation exif tags for creating thumbnails
+ * and displaying photos with exif data properly.
+ */
+fun normalizeExifOrientation(bitmap: Bitmap, bytesWithExif: ByteArray): Bitmap {
+    val orientation = ExifInterface(ByteArrayInputStream(bytesWithExif)).getAttributeInt(
+        ExifInterface.TAG_ORIENTATION,
+        ExifInterface.ORIENTATION_NORMAL
+    )
+    val matrix = Matrix()
+    when (orientation) {
+        ExifInterface.ORIENTATION_NORMAL -> return bitmap
+
+        ExifInterface.ORIENTATION_ROTATE_90 -> matrix.setRotate(90f)
+        ExifInterface.ORIENTATION_ROTATE_180 -> matrix.setRotate(180f)
+        ExifInterface.ORIENTATION_ROTATE_270 -> matrix.setRotate(-90f)
+
+        ExifInterface.ORIENTATION_FLIP_HORIZONTAL -> matrix.setScale(-1f, 1f)
+        ExifInterface.ORIENTATION_FLIP_VERTICAL -> {
+            matrix.setRotate(180f)
+            matrix.postScale(-1f, 1f)
+        }
+
+        ExifInterface.ORIENTATION_TRANSPOSE -> {
+            matrix.setRotate(90f)
+            matrix.postScale(-1f, 1f)
+        }
+        ExifInterface.ORIENTATION_TRANSVERSE -> {
+            matrix.setRotate(-90f)
+            matrix.postScale(-1f, 1f)
+        }
+        else -> return bitmap
+    }
+    return try {
+        val bmRotated: Bitmap = Bitmap.createBitmap(
+            bitmap,
+            0,
+            0,
+            bitmap.width,
+            bitmap.height,
+            matrix,
+            true
+        )
+        bitmap.recycle()
+        bmRotated
+    } catch (e: OutOfMemoryError) {
+        Timber.e(e)
+        bitmap
+    }
 }