+++ app/src/main/java/dev/leonlatsch/photok/ui/components/base/BaseActivity.kt
@@ -1,0 +1,55 @@
+/*
+ *   Copyright 2020-2021 Leon Latsch
+ *
+ *   Licensed under the Apache License, Version 2.0 (the "License");
+ *   you may not use this file except in compliance with the License.
+ *   You may obtain a copy of the License at
+ *
+ *        http://www.apache.org/licenses/LICENSE-2.0
+ *
+ *   Unless required by applicable law or agreed to in writing, software
+ *   distributed under the License is distributed on an "AS IS" BASIS,
+ *   WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
+ *   See the License for the specific language governing permissions and
+ *   limitations under the License.
+ */
+
+package dev.leonlatsch.photok.ui.components.base
+
+import android.os.Bundle
+import android.view.WindowManager
+import android.view.inputmethod.InputMethodManager
+import androidx.appcompat.app.AppCompatActivity
+import dev.leonlatsch.photok.R
+import dev.leonlatsch.photok.settings.Config
+
+/**
+ * Base for all activities.
+ *
+ * @since 1.0.0
+ * @author Leon Latsch
+ */
+abstract class BaseActivity : AppCompatActivity() {
+
+    override fun onCreate(savedInstanceState: Bundle?) {
+        setTheme(R.style.AppTheme)
+        super.onCreate(savedInstanceState)
+
+        if (!config.securityAllowScreenshots) {
+            window.addFlags(WindowManager.LayoutParams.FLAG_SECURE)
+        }
+    }
+
+    /**
+     * Abstract [Config], must be injected in implementations.
+     */
+    abstract var config: Config
+
+    /**
+     * Hide the soft-keyboard of displayed.
+     */
+    fun hideKeyboard() {
+        val inputManager = getSystemService(INPUT_METHOD_SERVICE) as InputMethodManager
+        inputManager.hideSoftInputFromWindow(currentFocus?.windowToken, 0)
+    }
+}