+++ app/src/main/java/dev/leonlatsch/photok/ui/process/base/BaseProcessBottomSheetDialogFragment.kt
@@ -1,5 +1,5 @@
 /*
- *   Copyright 2020 Leon Latsch
+ *   Copyright 2020-2021 Leon Latsch
  *
  *   Licensed under the Apache License, Version 2.0 (the "License");
  *   you may not use this file except in compliance with the License.
@@ -26,7 +26,7 @@ import dev.leonlatsch.photok.databinding.DialogBottomSheetProcessBinding
 import dev.leonlatsch.photok.other.hide
 import dev.leonlatsch.photok.other.show
 import dev.leonlatsch.photok.other.vanish
-import dev.leonlatsch.photok.ui.components.BindableBottomSheetDialogFragment
+import dev.leonlatsch.photok.ui.components.bindings.BindableBottomSheetDialogFragment
 
 /**
  * Abstract base for all process dialogs.