+++ app/src/main/java/dev/leonlatsch/photok/ui/settings/SettingsFragment.kt
@@ -1,5 +1,5 @@
 /*
- *   Copyright 2020 Leon Latsch
+ *   Copyright 2020-2021 Leon Latsch
  *
  *   Licensed under the Apache License, Version 2.0 (the "License");
  *   you may not use this file except in compliance with the License.
@@ -20,10 +20,13 @@ import android.app.Activity
 import android.content.Intent
 import android.net.Uri
 import android.os.Bundle
+import android.text.InputType
 import android.view.View
 import androidx.appcompat.widget.Toolbar
+import androidx.core.widget.addTextChangedListener
 import androidx.fragment.app.viewModels
 import androidx.navigation.fragment.findNavController
+import androidx.preference.EditTextPreference
 import androidx.preference.ListPreference
 import androidx.preference.Preference
 import androidx.preference.PreferenceFragmentCompat
@@ -31,10 +34,14 @@ import dagger.hilt.android.AndroidEntryPoint
 import dev.leonlatsch.photok.R
 import dev.leonlatsch.photok.databinding.BindingConverters
 import dev.leonlatsch.photok.other.setAppDesign
+import dev.leonlatsch.photok.other.show
 import dev.leonlatsch.photok.other.startActivityForResultAndIgnoreTimer
 import dev.leonlatsch.photok.settings.Config
 import dev.leonlatsch.photok.ui.components.Dialogs
 import dev.leonlatsch.photok.ui.process.BackupBottomSheetDialogFragment
+import dev.leonlatsch.photok.ui.settings.changepassword.ChangePasswordDialog
+import dev.leonlatsch.photok.ui.settings.hideapp.ToggleAppVisibilityDialog
+import javax.inject.Inject
 
 /**
  * Preference Fragment. Loads preferences from xml resource.
@@ -48,6 +55,9 @@ class SettingsFragment : PreferenceFragmentCompat() {
     private val viewModel: SettingsViewModel by viewModels()
     private var toolbar: Toolbar? = null
 
+    @Inject
+    lateinit var config: Config
+
     override fun onViewCreated(view: View, savedInstanceState: Bundle?) {
         super.onViewCreated(view, savedInstanceState)
 
@@ -60,14 +70,36 @@ class SettingsFragment : PreferenceFragmentCompat() {
     override fun onCreatePreferences(savedInstanceState: Bundle?, rootKey: String?) {
         setPreferencesFromResource(R.xml.settings, rootKey)
 
+        setupAppCategory()
+        setupSecurityCategory()
+        setupAdvancedCategory()
+        setupOtherCategory()
+    }
+
+
+    private fun setupAppCategory() {
         addCallbackTo<ListPreference>(Config.SYSTEM_DESIGN) {
             setAppDesign(it as String)
         }
+    }
 
+    private fun setupSecurityCategory() {
+        addActionTo(KEY_ACTION_CHANGE_PASSWORD) {
+            ChangePasswordDialog().show(childFragmentManager)
+        }
+
+        addActionTo(KEY_ACTION_HIDE_APP) {
+            ToggleAppVisibilityDialog().show(childFragmentManager)
+        }
+
+        configurePhoneDialPreference()
+    }
+
+    private fun setupAdvancedCategory() {
         addActionTo(KEY_ACTION_RESET) {
             Dialogs.showConfirmDialog(
                 requireContext(),
-                getString(R.string.settings_reset_confirmation)
+                getString(R.string.settings_advanced_reset_confirmation)
             ) { _, _ ->
                 viewModel.resetComponents()
             }
@@ -85,15 +117,9 @@ class SettingsFragment : PreferenceFragmentCompat() {
                 REQ_BACKUP
             )
         }
+    }
 
-        addActionTo(KEY_ACTION_CHANGE_PASSWORD) {
-            val dialog = ChangePasswordDialog()
-            dialog.show(
-                requireActivity().supportFragmentManager,
-                ChangePasswordDialog::class.qualifiedName
-            )
-        }
-
+    private fun setupOtherCategory() {
         addActionTo(KEY_ACTION_FEEDBACK) {
             val emailIntent = Intent(
                 Intent.ACTION_SENDTO,
@@ -125,8 +151,24 @@ class SettingsFragment : PreferenceFragmentCompat() {
             startActivity(intent)
         }
 
+        addActionTo(KEY_ACTION_CREDITS) {
+            findNavController().navigate(R.id.action_settingsFragment_to_creditsFragment)
+        }
+
         addActionTo(KEY_ACTION_ABOUT) {
             findNavController().navigate(R.id.action_settingsFragment_to_aboutFragment)
+        }
+    }
+
+    private fun configurePhoneDialPreference() {
+        val dialPreference = findPreference<EditTextPreference>(Config.SECURITY_DIAL_LAUNCH_CODE)
+        dialPreference?.setOnBindEditTextListener {
+            it.inputType = InputType.TYPE_CLASS_NUMBER or InputType.TYPE_NUMBER_FLAG_DECIMAL
+            it.addTextChangedListener { editable ->
+                if (editable?.length!! < 1) {
+                    it.setText(0.toString())
+                }
+            }
         }
     }
 
@@ -135,11 +177,7 @@ class SettingsFragment : PreferenceFragmentCompat() {
         if (requestCode == REQ_BACKUP && resultCode == Activity.RESULT_OK) {
             val uri = data?.data
             uri ?: return
-            val dialog = BackupBottomSheetDialogFragment(uri)
-            dialog.show(
-                requireActivity().supportFragmentManager,
-                BackupBottomSheetDialogFragment::class.qualifiedName
-            )
+            BackupBottomSheetDialogFragment(uri).show(requireActivity().supportFragmentManager)
         }
     }
 
@@ -167,9 +205,11 @@ class SettingsFragment : PreferenceFragmentCompat() {
 
         const val KEY_ACTION_RESET = "action_reset_safe"
         const val KEY_ACTION_CHANGE_PASSWORD = "action_change_password"
+        const val KEY_ACTION_HIDE_APP = "action_hide_app"
         const val KEY_ACTION_BACKUP = "action_backup_safe"
         const val KEY_ACTION_FEEDBACK = "action_feedback"
         const val KEY_ACTION_SOURCECODE = "action_sourcecode"
+        const val KEY_ACTION_CREDITS = "action_credits"
         const val KEY_ACTION_ABOUT = "action_about"
     }
 }